Fixing Dead Links
Fixed a few dead links throughout the website:

The FAQ had a dead link which was supposed to point to information about reciprocity. As Tom pointed out that this page is more public-facing (members likely wouldn't be on the FAQ), I changed this link to the NYS reciprocity website instead of the member-only page on reciprocity.

The home page had two dead links to Rensselaer County Public Safety (one on the logo, another under "Other Agencies." Swapped these out.

The DOH Resources page had two dead links as well, swapped those out.

Finally, the Rensselaer County Resources page had a dead link as well as one that didn't seem to match the correct description. Swapped that out for a new link and altered the description.